erlpiphany: Point APPNAME to ehal
The name of the module is "ehal" (ehal.beam), however when using the appname of "erlpiphany", init() would never find the module's priv dir if it was installed in erlang's path. This has worked for most people in the past because they would run the erlang code from the same directory as the ehal module (falling back to the non_existing match).